refactor(summary): split runtime facades
Keep text generation, summary quality, and summary backfill facades stable while moving formatting, grounding, query, dispatch, and runner logic into focused modules.

Facade wrappers keep monkeypatch seams for backfill dependencies and embed dispatch.
